02-todo/requirements.txt
asgiref==3.11.0
Django==4.0.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
pytz==2025.2
sqlparse==0.5.4
